Negamax works and is unbeatable.
Add depth information to the value of a terminal node. This makes the AI
chose the best path in the list of of available moves.

NOTE: Because I'm incrementing the depth each time I recurse, `min` must
be applied to the scores at each level, instead of `max`.  This is why
it's different from the algorithm listing on Wikipedia.

If I were decrementing the depth, then using `max` would be appropriate.